requirements.txt
aws-cdk-lib==2.56.0
constructs>=10.0.0,<11.0.0
boto3==1.26.34
requests==2.28.1
